requirements.txt
langchain_community
langchain_core
python-dotenv
langchain_text_splitters
sentence-transformers
langchain-chroma